Fix elevation data fetching and refactor to server action
